Moved PacketMetric to common.h
