+++ srtcore/group.h
@@ -29,96 +29,6 @@ Written by
     };
 #endif
 
-struct PacketMetric
-{
-    uint32_t pkts;
-    uint64_t bytes;
-
-    void update(uint64_t size)
-    {
-        ++pkts;
-        bytes += size;
-    }
-
-    void update(size_t mult, uint64_t value)
-    {
-        pkts += mult;
-        bytes += mult * value;
-    }
-
-    uint64_t fullBytes()
-    {
-        static const int PKT_HDR_SIZE = CPacket::HDR_SIZE + CPacket::UDP_HDR_SIZE;
-        return bytes + pkts * PKT_HDR_SIZE;
-    }
-};
-
-template <class METRIC_TYPE>
-struct MetricOp;
-
-template <class METRIC_TYPE>
-struct MetricUsage
-{
-    METRIC_TYPE local;
-    METRIC_TYPE total;
-
-    void Clear()
-    {
-        MetricOp<METRIC_TYPE>::Clear(local);
-    }
-
-    void Init()
-    {
-        MetricOp<METRIC_TYPE>::Clear(total);
-        Clear();
-    }
-
-    void Update(uint64_t value)
-    {
-        local += value;
-        total += value;
-    }
-
-    void UpdateTimes(size_t mult, uint64_t value)
-    {
-        local += mult * value;
-        total += mult * value;
-    }
-};
-
-template <>
-inline void MetricUsage<PacketMetric>::Update(uint64_t value)
-{
-    local.update(value);
-    total.update(value);
-}
-
-template <>
-inline void MetricUsage<PacketMetric>::UpdateTimes(size_t mult, uint64_t value)
-{
-    local.update(mult, value);
-    total.update(mult, value);
-}
-
-template <class METRIC_TYPE>
-struct MetricOp
-{
-    static void Clear(METRIC_TYPE& m)
-    {
-        m = 0;
-    }
-};
-
-template <>
-struct MetricOp<PacketMetric>
-{
-    static void Clear(PacketMetric& p)
-    {
-        p.pkts = 0;
-        p.bytes = 0;
-    }
-};
-
 class CUDTGroup
 {
     friend class CUDTUnited;
